Move search into own component, start of WeatherForecast component
