+++ src/App.js
@@ -1,6 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import './styles/main.css';
 import Header from './Components/Header/Header';
+import WeatherForecastSearch from './Components/WeatherForecastSearch/WeatherForecastSearch';
 import WeatherForecast from './Components/WeatherForecast/WeatherForecast';
 import fetchWeatherForecast from './Services/weatherApi';
 
@@ -8,23 +9,29 @@ const App = () => {
   // eslint-disable-next-line no-unused-vars
   const [searchedLocation, setSearchedLocation] = useState(null);
   const [searchedLatLong, setSearchedLatLong] = useState(null);
+  // eslint-disable-next-line no-unused-vars
   const [weatherForecast, setWeatherForecast] = useState(null);
+
 
   useEffect(() => {
     if (searchedLatLong == null) {
       return;
     }
-    fetchWeatherForecast(searchedLatLong).then(data => setWeatherForecast(data))y
+    fetchWeatherForecast(searchedLatLong).then((data) =>
+      setWeatherForecast(data)
+    );
   }, [searchedLatLong]);
 
   return (
     <>
       <Header />
-      <WeatherForecast
+      <WeatherForecastSearch
         setSearchedLocation={setSearchedLocation}
         setSearchedLatLong={setSearchedLatLong}
-        weatherForecast={weatherForecast}
         setWeatherForecast={setWeatherForecast}
+      />
+      <WeatherForecast
+        weatherForecast={weatherForecast}
       />
     </>
   );
